Factor out store tests to test_store.py
